services/backend/pyproject.toml
[tool.poetry]
name = "kakeibo-journal"
version = "0.1.0"
description = ""
authors = ["Laurens Sandt <laurens.sandt@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "*"
fastapi = "*"
tortoise-orm = "*"
aerich = "*"
asyncpg = "*"
bcrypt = "*"
passlib = "*"
python-multipart = "*"
python-jose = "*"


[tool.poetry.group.dev.dependencies]
httpx = "*"
pytest = "*"
ruff = "*"
mypy = "*"

[tool.ruff.lint]
select =  [
    # isort
    "I",
    ]

[tool.ruff.format]
quote-style = "single"


[tool.aerich]
tortoise_orm = "src.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
